feat: Bonus feature 3 added. test files for both logic units and some fixes in services completed.

+++ src/main/java/com/anujl/online_quiz_application/service/QuizService.java
@@ -101,7 +101,7 @@ throw new ResourceNotFoundException("No Quiz in the DB");
                             correct++;
                         }
                     }
-                    score+=entity.getPoints()*(correct/totalCorrectOptions);
+                    score+= (int) (entity.getPoints()*((double)correct/totalCorrectOptions));
                 }
                 default -> throw new IllegalArgumentException("Unknown question type.");
             }
